Migrate project to Maven
