+++ CHANGELIST.md
@@ -1,0 +1,7 @@
+1. Migrate to Maven
+    - Changes
+        - Use maven to manage dependencies and build the project
+        - Use Java 17
+        - Update .gitignore
+    - Reasoning
+        - Maven is a widely used build tool for Java projects. It is easy to build and test the project using Maven.
