+++ README.md
@@ -5,17 +5,21 @@ Rewrite and improve the code after your own liking.
 
 Think: you are responsible for the solution, this is a solution you will have to put your name on.
 
-
 ## Handing in the assignment
 
-Reason how you have been thinking and the decisions you took. 
+Reason how you have been thinking and the decisions you took.
 You can hand in the result any way you feel (git patch, pull-request or ZIP-file).
 Note: the Git history must be included.
 
+## How to run the application
 
-## To run the test:
+1. Build the application
 
-```
-javac src/*.java
-java -cp src Main
-```
+    ```
+    mvn clean package
+    ```
+2. Run the application
+
+    ```
+    java -jar target/refactoring-java-1.0-SNAPSHOT.jar
+    ```
