+++ src/main/java/com/mithwick93/refactoring/java/Customer.java
@@ -1,3 +1,5 @@
+package com.mithwick93.refactoring.java;
+
 import java.util.List;
 
 public class Customer {
